Fixed connection problems when using multiple controller nodes
