+++ eiscp-in.html
@@ -1,0 +1,33 @@
+<script type="text/x-red" data-template-name="eiscp-in">
+    <div class="form-row">
+        <label for="node-input-name"><i class="icon-tag"></i> Name</label>
+        <input type="text" id="node-input-name" placeholder="Name">
+    </div>
+    	<div class="form-row">
+        <label for="node-input-controller"><i class="icon-bookmark"></i> Controller</label>
+        <input type="text" id="node-input-controller">
+    </div>
+
+
+</script>
+
+<script type="text/x-red" data-help-name="eiscp-in">
+    <p>Use this to <b>inject</b> flows from EISCP device<br/>
+</script>
+
+<script type="text/javascript">
+    RED.nodes.registerType('eiscp-in', {
+        category: 'input',
+        color: '#26b050',
+        defaults: {
+            name: {value: ""},
+            controller: {value: "", type: "eiscp-controller"}
+        },
+        inputs: 0,
+        outputs: 1,
+        icon: "bridge-dash.png",
+        label: function () {
+            return (this.groupaddr || this.name || "eiscp");
+        }
+    });
+</script>
