--- eiscp.html
@@ -1,109 +1,0 @@
-<script type="text/x-red" data-template-name="eiscp-controller">
-    <div class="form-row">
-        <label for="node-config-input-name"><i class="icon-tag"></i> Name</label>
-        <input type="text" id="node-config-input-name" placeholder="Name">
-    </div>
-    <div class="form-row">
-        <label for="node-config-input-host"><i class="icon-bookmark"></i> Host</label>
-        <input type="text" id="node-config-input-host">
-    </div>
-    <div class="form-row">
-        <label for="node-config-input-port"><i class="icon-bookmark"></i> Port</label>
-        <input type="text" id="node-config-input-port" placeholder="60128">
-    </div>
-    <div class="form-row">
-        <label for="node-config-input-model"><i class="icon-bookmark"></i> Model</label>
-        <input type="text" id="node-config-input-model" placeholder="model (for example: DTR-6.6)">
-    </div>
-</script>
-
-<script type="text/x-red" data-template-name="eiscp-out">
-    <div class="form-row">
-        <label for="node-input-name"><i class="icon-tag"></i> Name</label>
-        <input type="text" id="node-input-name" placeholder="Name">
-    </div>
-    <div class="form-row">
-        <label for="node-input-controller"><i class="icon-bookmark"></i> Controller</label>
-        <input type="text" id="node-input-controller">
-    </div>
-</script>
-
-<script type="text/x-red" data-template-name="eiscp-in">
-    <div class="form-row">
-        <label for="node-input-name"><i class="icon-tag"></i> Name</label>
-        <input type="text" id="node-input-name" placeholder="Name">
-    </div>
-    	<div class="form-row">
-        <label for="node-input-controller"><i class="icon-bookmark"></i> Controller</label>
-        <input type="text" id="node-input-controller">
-    </div>
-
-
-</script>
-
-<script type="text/x-red" data-help-name="eiscp-out">
-    <p>
-        Use this to <b>send</b> EISCP packets to Onkyo/Integra device familiar with EISCP protocol.<br/>
-    	<b>msg.payload</b> must be a JavaScript object (with <b>cmd<b> and <b>args</b> fields) or a string.
-    </p>
-    <p>
-        For example: <code>system-power=on</code> or <code>system-power=standby</code>
-    </p>
-    <p>
-        Please, take a look at full list of commands here: https://raw.githubusercontent.com/estbeetoo/eiscp.js/master/eiscp-commands.json
-    <p>
-</script>
-
-<script type="text/x-red" data-help-name="eiscp-in">
-    <p>Use this to <b>inject</b> flows from EISCP device<br/>
-</script>
-
-<script type="text/javascript">
-    RED.nodes.registerType('eiscp-controller', {
-        category: 'config',
-        defaults: {
-            name: {value: ""},
-            host: {value: "127.0.0.1", required: true},
-            port: {value: 60128, required: false, validate: RED.validators.number()},
-            model: {value: "", required: true}
-        },
-        label: function () {
-            return (this.name || 'eiscp' ) + "@" + this.host + ":" + this.port;
-        }
-    });
-</script>
-
-<script type="text/javascript">
-    RED.nodes.registerType('eiscp-out', {
-        category: 'output',
-        color: '#26b050',
-        defaults: {
-            name: {value: ""},
-            controller: {value: "", type: "eiscp-controller"}
-        },
-        inputs: 1,
-        outputs: 0,
-        align: 'right',
-        icon: "bridge-dash.png",
-        label: function () {
-            return (this.groupaddr || this.name || "eiscp");
-        }
-    });
-</script>
-
-<script type="text/javascript">
-    RED.nodes.registerType('eiscp-in', {
-        category: 'input',
-        color: '#26b050',
-        defaults: {
-            name: {value: ""},
-            controller: {value: "", type: "eiscp-controller"}
-        },
-        inputs: 0,
-        outputs: 1,
-        icon: "bridge-dash.png",
-        label: function () {
-            return (this.groupaddr || this.name || "eiscp");
-        }
-    });
-</script>
