+++ package.json
@@ -2,11 +2,10 @@
   "name": "node-red-contrib-eiscp",
   "version": "0.3.2",
   "description": "Integra / Onkyo (ISCP protocol) for node-red",
-  "main": "eiscp.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "repository": {
+  "_repository": {
     "type": "git",
     "url": "git+https://github.com/estbeetoo/node-red-contrib-eiscp.git"
   },
@@ -20,7 +19,9 @@
   ],
   "node-red": {
     "nodes": {
-      "eiscp": "eiscp.js"
+      "eiscp-controller": "eiscp-controller.js",
+      "eiscp-in": "eiscp-in.js",
+      "eiscp-out": "eiscp-out.js"
     }
   },
   "author": "Alexander Borovsky <aborovsky@beetoo.me> (http://beetoo.me)",
